+++ CHANGES.md
@@ -1,5 +1,8 @@
 ## Unchanged
 
+#### Fixed
+
+  - `after` and `enrich` interceptors now no longer run if there is no `db` effect, rather than running against the `db` coffect. [#453](https://github.com/Day8/re-frame/issues/453)
 
 
 ## 0.10.6 (2018-09-03)